KIALI-1726: Move Logo images into their own config file 'logos.ts'
