+++ src/components/About/AboutUIModal.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { AboutModal, Icon } from 'patternfly-react';
 import { Component } from '../../store/Store';
-import { config, KialiLogo } from '../../config';
+import { config } from '../../config';
+import { KialiLogo } from '../../logos';
 
 const KIALI_CORE_COMMIT_HASH = 'Kiali core commit hash';
 const KIALI_CORE_VERSION = 'Kiali core version';
