+++ src/logos.ts
@@ -1,0 +1,7 @@
+// All logo images go in here
+
+/** Istio logo */
+export const IstioLogo = require('./assets/img/istio-logo.svg');
+
+/** Kiali logo */
+export const KialiLogo = require('./assets/img/logo-alt.svg');
